add exp module docs
