+++ oneflow/python/nn/modules/array_ops.py
@@ -41,8 +41,12 @@ from oneflow.python.nn.common_types import _size_1_t, _size_2_t, _size_3_t
 from oneflow.python.ops.nn_ops import calc_pool_padding, get_dhw_offset
 import oneflow.python.framework.id_util as id_util
 from oneflow.python.framework.tensor import register_tensor_op_by_module
+from oneflow.python.framework.tensor import register_op_by_module
 
 
+@oneflow_export("Transpose")
+@register_tensor_op_by_module("tmp.transpose")
+@register_op_by_module("tmp.transpose")
 class Transpose(Module):
     r"""
     """
@@ -71,23 +75,3 @@ class Transpose(Module):
 
     def forward(self, x):
         return self._op(x)[0]
-
-
-@oneflow_export("tmp.transpose")
-def transpose(
-    a,
-    perm: Sequence[int] = None,
-    conjugate: bool = False,
-    batch_axis_non_change: bool = False,
-    name: Optional[str] = None,
-):
-    return Transpose(perm, conjugate, batch_axis_non_change, name)(a)
-
-
-if __name__ == "__main__":
-    import numpy as np
-
-    flow.enable_eager_execution(True)
-    x = flow.Tensor(np.random.randn(1, 2, 3))
-    y = flow.tmp.transpose(x, perm=[2, 0, 1])
-    print(y.shape)
