+++ oneflow/python/nn/modules/math_ops.py
@@ -608,3 +608,34 @@ class Add(Module):
             return ScalarAddByTensor()(x, y)
         else:
             return BroadcastAdd()(x, y)
+
+
+@oneflow_export("Exp")
+@register_tensor_op_by_module("exp")
+@register_op_by_module("exp")
+class Exp(Module):
+    r"""
+    Returns a new tensor with the exp of the elements of :attr:`x`.
+    .. math::
+        \text{y}_{i} = \exp(\text{x}_{i})
+    Args:
+        x (flow.Tensor): X.
+    
+    For example: 
+    .. code-block:: python 
+        import numpy as np
+        import oneflow as flow
+
+        x = flow.Tensor(np.array([1, 2, 3]).astype(np.float32))
+        y = x.exp()
+
+        y [ 2.7182817  7.389056  20.085537 ]
+
+    """
+
+    def __init__(self) -> None:
+        super().__init__()
+        self._op = flow.builtin_op("exp").Input("x").Output("y").Build()
+
+    def forward(self, x):
+        return self._op(x)[0]
